+++ tests/steps_file.js
@@ -4,24 +4,23 @@
 
 module.exports = function() {
   return actor({
+        updateNotifications:function(){
+            bfUrl = this.getUrl()
+        },
         login:function(){
 
-            // user = process.env.dlbs_user;
-            // password = process.env.dlbs_passwd;
+            var user = process.env.dlbs_user;
+            var password = process.env.dlbs_passwd;
 
-            var b64string = "RTJEdFMhem96akkhMUYmRjZzTUdj";
-            var buf = Buffer.from(b64string, 'base64'); // Ta-da
-
-            var user = "dlbs-plugin-bot";
-            var password = buf.toString();
-
-            this.amOnPage('/');
+            this.amOnPage('https://www.dealabs.com/');
             this.click("#login_header a");
             this.wait(1);
             this.fillField("username", user);
             this.fillField("password", password);
             this.click(".validate_div .validate_button_form");
-            this.saveScreenshot('connexion.png');
+            this.wait(1);
+            this.dontSee("Mot de passe incorrect.");
+            // this.saveScreenshot('connexion.png');
         }
     // Define custom steps here, use 'this' to access default methods of I.
     // It is recommended to place a general 'login' function here.